% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_brain_regions.R
\name{list_brain_regions}
\alias{list_brain_regions}
\title{List all brain region names in AAL/BA system}
\usage{
list_brain_regions(template = c("aal", "ba"))
}
\arguments{
\item{template}{A character value which indicates the templates of interest
(\code{"aal"} or \code{"ba"}). Use both of them by default.}
}
\value{
Return a list of charcter vectors.
Each character vector is all the brain region names defined by one of the template.
}
\description{
Input template names, output all the brain region names.
}
\examples{
# List all brain region names defined by AAL.
list_brain_regions(template = "aal")
}
