% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_to_region_name.R
\name{mni_to_region_name}
\alias{mni_to_region_name}
\title{MRI-labeling: MNI to AAL}
\usage{
mni_to_region_name(x, y, z, distance = T)
}
\arguments{
\item{x}{A number}

\item{y}{A number}

\item{z}{A number}

\item{distance}{A logical value which indicates whether the distance-to-nearest-aal-region information should be shown (default=T). This could be turned off to speed up labeling process.}
}
\value{
If distance mode is on (distance=T), output a list with brain region name along with the corresponding distance.If distance=F, output a string of region name when available, otherwise output "Not exactly correspond to aal-labeled brain region. Please set distance=T if you want the nearest aal-labeled region name.".
}
\description{
Input an MNI coordinate, output the corresponding AAL (Automated Anatomical Labeling) brain region name.
}
\examples{
# aal-corresponding point with distance mode on
mni_to_region_name(26,0,0,distance=T)
# aal-corresponding point with distance mode off (much faster)
mni_to_region_name(26,0,0,distance=F)

#non-aal-corresponding point with distance mode on (output the nearest aal region name)
mni_to_region_name(0,0,0,distance=T)
#non-aal-corresponding point with distance mode off (output none aal region name)
mni_to_region_name(0,0,0,distance=F)
}
\seealso{
\code{\link[utils]{head}}
}
