% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_name_to_mni.R
\name{region_name_to_mni}
\alias{region_name_to_mni}
\title{Output the MNI coordinates of brain regions defined by AAL/BA system}
\usage{
region_name_to_mni(region_names, template = "aal")
}
\arguments{
\item{region_names}{A character vector which indeicates the brain region names of interest.
Use \code{list_brain_regions()} to see all brain region names defined by AAL/BA system.}

\item{template}{One character value which indicates the templates to use
(\code{"aal"} or \code{"ba"}). Use \code{"aal"} by default.}
}
\value{
Return a list of data frames and each of them correspond to a template.
Each data frame consists the MNI coordinates of the input brain region.
}
\description{
Input a brain region name, output the corresponding MNI coordinates.
This is the inverse function of the function \code{mni_to_region_name()}.
}
\examples{
# Get the MNI cooridnates of the right precentral region
# defined by AAL template
region_name_to_mni(region_names = "Precentral_R", template = "aal")

# Get the MNI cooridnates of both the right and left precentral region
# defined by AAL template
region_name_to_mni(
  region_names = c("Precentral_R", "Precentral_L"),
  template = "aal"
)
}
\seealso{
\code{\link{mni_to_region_name}} \cr
\code{\link{list_brain_regions}}
}
