% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_to_region_name.R
\name{mni_to_region_name}
\alias{mni_to_region_name}
\title{MRI-labeling: label the brain MNI coordinate by AAL/BA system}
\usage{
mni_to_region_name(x, y, z, distance = T, template = c("aal", "ba"))
}
\arguments{
\item{x}{The numeric x value of the MNI coordinate.}

\item{y}{The numeric y value of the MNI coordinate.}

\item{z}{The numeric z value of the MNI coordinate.}

\item{distance}{A logical value which indicates whether the closest region
should be shown when there is no exact match (\code{default = T}).
This could be turned off to speed up the labeling process.}

\item{template}{A character vector which indicates the templates to use,
`aal` by default.}
}
\value{
If distance mode is on (\code{distance = T}),
output a list of brain region names along with the corresponding distances (mm).
Please set \code{distance = T} if you want the closest region name even when
there is no exact matching brain region.
If distance mode is off (\code{distance = F}),
output a string of region names only when available,
otherwise output 'NULL'.
}
\description{
Input an MNI coordinate, output the corresponding AAL/BA brain region name.
}
\examples{
# Exact matching brain region with distance mode on
mni_to_region_name(26, 0, 0, distance = T)
# Exact matching brain region with distance mode off (much faster)
mni_to_region_name(26, 0, 0, distance = F)

# No exact matching brain region with distance mode on (output the nearest brain region name)
mni_to_region_name(0, 0, 0, distance = T)
# No exact matching brain region  with distance mode off (output nothing)
mni_to_region_name(0, 0, 0, distance = F)

# Only acquire AAL region name
mni_to_region_name(26, 0, 0, distance = T, template = "aal")
}
\seealso{
\code{\link{region_name_to_mni}}
}
