% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cluster_composition.R
\name{show_cluster_composition}
\alias{show_cluster_composition}
\title{Show brain cluster composition}
\usage{
show_cluster_composition(coordinate_matrix, template = c("aal", "ba"))
}
\arguments{
\item{template}{One character vector which indicates the templates to use
(\code{"aal"} or \code{"ba"}). Use \code{"aal"} by default.}
}
\value{
Return a list of data frames and each of them correspond to a template.
Each data frame consists the brain region names and the corresponding percentage
of the input brain cluster.
}
\description{
Input a matrix of all MNI coordinates within a cluster,
and output the brain regions and the percentage of each region.
}
\seealso{
\code{\link{show_nii_clusters}}
}
